Move task management into separate object
